+++ zkp/golang/integration-test/e2e_test.go
@@ -26,11 +26,11 @@ import (
 	"time"
 
 	"github.com/hyperledger-labs/zeto/internal/testutils"
-	"github.com/hyperledger-labs/zeto/pkg/core"
-	"github.com/hyperledger-labs/zeto/pkg/node"
-	"github.com/hyperledger-labs/zeto/pkg/smt"
-	"github.com/hyperledger-labs/zeto/pkg/storage"
-	"github.com/hyperledger-labs/zeto/pkg/utxo"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/core"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/node"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/smt"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/storage"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/utxo"
 	"github.com/iden3/go-iden3-crypto/babyjub"
 	"github.com/iden3/go-iden3-crypto/poseidon"
 	"github.com/iden3/go-rapidsnark/prover"
