+++ zkp/golang/internal/sparse-merkle-tree/node/errors.go
