+++ zkp/golang/internal/sparse-merkle-tree/smt/errors.go
