+++ zkp/golang/internal/sparse-merkle-tree/smt/merkletree.go
@@ -20,10 +20,10 @@ import (
 	"math/big"
 	"sync"
 
-	"github.com/hyperledger-labs/zeto/internal/node"
-	"github.com/hyperledger-labs/zeto/internal/storage"
-	"github.com/hyperledger-labs/zeto/internal/utils"
-	"github.com/hyperledger-labs/zeto/pkg/core"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/node"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/storage"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/utils"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/core"
 )
 
 // MAX_TREE_HEIGHT is the maximum number of levels of the sparse merkle tree.
