+++ zkp/golang/internal/sparse-merkle-tree/storage/errors.go
