+++ zkp/golang/internal/sparse-merkle-tree/utxo/fungible_test.go
