+++ zkp/golang/internal/sparse-merkle-tree/utxo/nonfungible_test.go
