+++ zkp/golang/pkg/sparse-merkle-tree/core/merkletree.go
