+++ zkp/golang/pkg/sparse-merkle-tree/core/node.go
