+++ zkp/golang/pkg/sparse-merkle-tree/core/proof.go
