+++ zkp/golang/pkg/sparse-merkle-tree/core/storage.go
