+++ zkp/golang/pkg/sparse-merkle-tree/node/node.go
@@ -19,8 +19,8 @@ package node
 import (
 	"math/big"
 
-	"github.com/hyperledger-labs/zeto/internal/node"
-	"github.com/hyperledger-labs/zeto/pkg/core"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/node"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/core"
 )
 
 func NewEmptyNode() core.Node {
