+++ zkp/golang/pkg/sparse-merkle-tree/smt/merkletree.go
@@ -14,22 +14,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package smt
 
 import (
-	"github.com/hyperledger-labs/zeto/pkg/core"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/smt"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/core"
 )
 
-type IndexOnly struct {
-	Index core.NodeIndex
-}
-
-func NewIndexOnly(index core.NodeIndex) *IndexOnly {
-	return &IndexOnly{
-		Index: index,
-	}
-}
-
-func (f *IndexOnly) CalculateIndex() (core.NodeIndex, error) {
-	return f.Index, nil
+func NewMerkleTree(db core.Storage, maxLevels int) (core.SparseMerkleTree, error) {
+	return smt.NewMerkleTree(db, maxLevels)
 }
