+++ zkp/golang/pkg/sparse-merkle-tree/utxo/utxo.go
@@ -19,9 +19,9 @@ package utxo
 import (
 	"math/big"
 
-	"github.com/hyperledger-labs/zeto/internal/utils"
-	"github.com/hyperledger-labs/zeto/internal/utxo"
-	"github.com/hyperledger-labs/zeto/pkg/core"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/utils"
+	"github.com/hyperledger-labs/zeto/internal/sparse-merkle-tree/utxo"
+	"github.com/hyperledger-labs/zeto/pkg/sparse-merkle-tree/core"
 	"github.com/iden3/go-iden3-crypto/babyjub"
 )
